+++ sensors.h
@@ -18,8 +18,8 @@
 #define ET_CORR_BINS  5
 
 // Read raw ADC and convert to engineering units
-uint16_t readTPS();    // returns 0–1000 per-mille (0 = 0%, 1000 = 100%)
-float    readFPS();    // returns bar
+uint8_t  readTPS();    // returns 0–100 percent
+uint8_t  readFPS();    // returns 0–80, units = 0.125 bar (1/8 bar per count)
 int16_t  readIAT();    // returns whole °C
 int16_t  readET();     // returns whole °C
-float    readBatV();   // returns battery voltage in V (voltage divider ratio 3.2:1)
+uint8_t  readBatV();   // returns battery voltage in 1/16 V units (0.0625 V per count)
